Corrected authorship designation for 2008-2011 user contribution
Adjusted in accordance with the author's rights as stipulated by the
German Act on Copyright and Related Rights (Urheberrechtsgesetz),
where the code was authored.

As section 13 of the Urheberrechtsgesetz states:

> The author has the right to be identified as the author of the work.
> He may determine whether the work shall bear a designation of
> authorship and which designation is to be used.

Copyright and authorship rights are non-transferrable under
Section 29(1), and the author furthermore certifies never to have
entered into a Contributor Agreement or equivalent with any parties
associated with OpenFOAM  or otherwise relinquished the right to
assert authorship rights.